ui/requirements.txt
# Core dependencies
gradio>=4.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.26.0
opencv-python>=4.8.0
Pillow>=10.0.0
matplotlib>=3.0,<4.0

# Optional: for better performance
albumentations>=1.3.0

# Optional: for YAML config
pyyaml>=6.0

